Reformatted code into seperate files for better readibility
