+++ analysis/voxelWise/main.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.insert(0, '../')
 
-import os
+import os, timeit
 import numpy as np
 import model_utils
 import visual
@@ -16,8 +16,8 @@ if not os.path.exists(model_dir):
     os.makedirs(model_dir)
 
 # Path to save the model to
-model_name = 'test_cv.pkl'
-model_path = os.path.join(model_dir, model_name)
+model_name = 'test_cv'
+model_path = os.path.join(model_dir, model_name + '.pkl')
 if os.path.isfile(model_path):
     # file exists
     print('This model already exists: ', model_path)
@@ -27,22 +27,21 @@ if os.path.isfile(model_path):
 
 notification_system = NotificationSystem()
 
-# ct_sequences = ['wcoreg_RAPID_TMax_[s]', 'wcoreg_RAPID_MTT_[s]', 'wcoreg_RAPID_CBV', 'wcoreg_RAPID_CBF']
-# ct_sequences = ['wcoreg_RAPID_Tmax', 'wcoreg_RAPID_MTT', 'wcoreg_RAPID_rCBV', 'wcoreg_RAPID_rCBF']
-# ct_sequences = ['wcoreg_RAPID_TMax_[s]']
-# mri_sequences = ['wcoreg_VOI_lesion']
-
 IN, OUT = data_loader.load_saved_data(data_dir)
 # IN, OUT = manual_data.load(data_dir)
 
+
 rf = 0
 rf_dim = [rf, rf, rf]
+print('Evaluating', model_name, 'with rf:', rf_dim)
 
-# model_utils.patient_wise_kfold_data_split(data_dir, IN, OUT, rf_dim)
-# model_utils.train_CV(data_dir)
+save_dir = data_dir = os.path.join(data_dir, 'cv_data')
 
-score, roc_auc, f1 = model_utils.evaluate_model(data_dir, model_dir, model_name, IN, OUT, rf_dim)
+start = timeit.default_timer()
+score, roc_auc, f1 = model_utils.evaluate_crossValidation(save_dir, model_dir, model_name, IN, OUT, rf_dim)
+elapsed = timeit.default_timer() - start
+print('Evaluation done in: ', elapsed)
 
-# title = model_name + ' finished Cross-Validation'
-# body = 'accuracy ' + str(score) + '\n' + 'ROC AUC ' + str(roc_auc) + '\n' + 'F1 ' + str(f1)
-# notification_system.send_message(title, body)
+title = model_name + ' finished Cross-Validation'
+body = 'accuracy ' + str(score) + '\n' + 'ROC AUC ' + str(roc_auc) + '\n' + 'F1 ' + str(f1) + '\n' + 'RF ' + str(rf) + '\n' + 'Time elapsed ' + str(elapsed)
+notification_system.send_message(title, body)
